requirements.txt
Django==6.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9